requirements.txt
# Framework del Bot y utilidades
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
Requests==2.32.5
httpx==0.27.0

# Inteligencia Artificial (Groq)
groq==0.9.0

# Audio (Whisper y TTS)
openai_whisper==20250625
edge_tts==7.2.3
pydub==0.25.1
torch==2.8.0

# Base de Datos (Firebase)
firebase-admin==6.5.0

# Imágenes
Pillow==10.4.0

# Análisis de Sentimiento y Traducción
transformers==4.43.3
pysentimiento==0.7.3 
langdetect==1.0.9